pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "testscript_eval"
description = "TestScript FHIR resource evaluator"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["fhir"]
authors = [{ name = "beda.software", email = "ilya@beda.software" }]
dependencies = ["requests>=2.25.1", "aiohttp>=3.6.3", "pytz", "aiofiles", "fhirpathpy", "pyyaml-include", "jsonpath_ng", "funcy", "fhirpy", "pytest_asyncio"]
version = "0.0.1"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"

[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
(
  /(
    | \.git
    | \.pytest_cache
    | htmlcov
    | locales
    | resources
    | requirements
    | embed
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 100
default_section = "THIRDPARTY"
known_first_party = ["testscript_eval"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --color=yes --disable-warnings"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"

[tool.setuptools]
py-modules = ["testscript_eval"]

[project.entry-points.pytest11]
testscript-eval-py  = "testscript_eval.pytest_plugin"
